Add files via upload
This analysis provides a predictive classification problem with the y-variable being CROSS_SELL_SUCCESS. The aim is to increase the AUC score to understand which features have the largest impact on the success of various cross sales. 

The dataset is from a small San Francisco based startup active in the food industry.